server/requirements.txt
requests
Django
Pillow
gunicorn
python-dotenv
djangorestframework